Mobile stuff
Create Ticket Page done (properly) and some navbar fixes.
