+++ frontend/src/components/CreateTicket/CreateTicketComp.tsx
@@ -22,6 +22,7 @@ import { getFaultTypes, CreatTicket } from "@/services/tickets.service";
 import { MapboxContextProps } from "@/context/MapboxContext";
 import { cn } from "@/lib/utils";
 import { Locate, Pin, PinOff, Info } from "lucide-react";
+import { FiMapPin, FiArrowRight, FiMap } from "react-icons/fi";
 import { PlaceKit, PlaceKitOptions } from "@placekit/autocomplete-react";
 import CustomMarker from "../../../public/customMarker.svg";
 import { PKResult } from "@placekit/client-js";
@@ -80,6 +81,7 @@ const CreateTicketComp: React.FC<Props> = ({ className, useMapboxProp }) => {
   const [isPinDropped, setIsPinDropped] = useState(false);
   const [tooltipVisible, setTooltipVisible] = useState(true);
   const [isClient, setIsClient] = useState(false);
+  const [isMapOpen, setIsMapOpen] = useState(false); // For mobile map modal
 
   const memoizedApiKey = useMemo(
     () => String(process.env.PLACEKIT_API_KEY),
@@ -231,22 +233,271 @@ const CreateTicketComp: React.FC<Props> = ({ className, useMapboxProp }) => {
     return selectedSeverity === severity ? "border-blue-500 border-3" : "";
   };
 
-  
   return (
     <div className="flex justify-center items-center h-full w-full px-4 overflow-hidden">
       <ToastContainer />
 
-      <div className="flex w-full max-w-screen-xl h-[40rem] rounded-lg overflow-hidden">
-        {/* Form Section */}
-        <div className="w-1/2 p-6 bg-white flex flex-col justify-center overflow-y-auto">
+      {/* Desktop View */}
+<div className="hidden sm:flex w-full max-w-screen-xl h-[40rem] rounded-lg overflow-hidden">
+  {/* Form Section */}
+  <div className="w-1/2 p-6 bg-white flex flex-col overflow-hidden">
+    <h2 className="text-2xl text-center font-bold mt-2 mb-6">
+      Report a Fault
+    </h2>
+    <div className="flex-1 overflow-y-auto">
+      <form
+        ref={formRef}
+        onSubmit={handleSubmit}
+        className="flex flex-col gap-y-8"
+      >
+        {/* Fault Type */}
+        {faultTypes.length > 0 ? (
+          <Autocomplete
+            label={
+              <span className="font-semibold text-sm">
+                Fault type <sup className="text-blue-500">*</sup>
+              </span>
+            }
+            labelPlacement="outside"
+            name="fault-type"
+            placeholder="Fault Type"
+            fullWidth
+            defaultItems={faultTypes}
+            disableSelectorIconRotation
+            isClearable={false}
+            menuTrigger={"input"}
+            size={"lg"}
+            type="text"
+            autoComplete="new-fault"
+            onSelectionChange={(key) => setSelectedFault(key as string)}
+          >
+            {(faultType) => (
+              <AutocompleteItem
+                key={faultType.asset_id}
+                textValue={faultType.asset_id}
+              >
+                <div className="flex gap-2 items-center">
+                  <img
+                    src={faultType.assetIcon}
+                    alt={faultType.asset_id}
+                    className="flex-shrink-0 w-6 h-6"
+                  />
+                  <span className="text-small">{faultType.asset_id}</span>
+                </div>
+              </AutocompleteItem>
+            )}
+          </Autocomplete>
+        ) : (
+          <p>Loading fault types...</p>
+        )}
+
+        {/* Description */}
+        <Textarea
+          label={<span className="font-semibold text-sm">Description</span>}
+          labelPlacement="outside"
+          name="fault-description"
+          placeholder="Add Description..."
+          onChange={(e) => setFaultDescription(e.target.value)}
+        />
+
+        {/* Address */}
+        <div className="w-full">
+          <span className="font-semibold text-sm">Address:</span>
+          <PlaceKit
+            apiKey={memoizedApiKey}
+            options={pkaOptions}
+            className="w-full"
+            onPick={handleSuggestionPick}
+            placeholder="Search for an address..."
+          />
+
+          <div>
+            <div className="flex flex-col gap-y-0.5 text-xs ps-2">
+              <span>{selectedAddress?.street?.name}</span>
+              <span>{selectedAddress?.county}</span>
+              <span>{selectedAddress?.city}</span>
+              <span>{selectedAddress?.administrative}</span>
+            </div>
+          </div>
+        </div>
+
+        {/* Fault Severity */}
+        <div>
+          <span className="font-semibold text-sm">Fault Severity</span>
+          <div className="flex h-[3.5rem] justify-center">
+            <ButtonGroup aria-label="Basic example" className="flex h-full">
+              <Button
+                variant="bordered"
+                className={`h-full ${getButtonStyle("Minor")}`}
+                onClick={() => handleSelect("Minor")}
+              >
+                <div className="flex flex-col min-w-32 px-2 font-sm rounded-3xl justify-center items-center h-full">
+                  <img
+                    width="25"
+                    height="auto"
+                    src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/fault_icon_minor.webp"
+                    alt="Minor"
+                  />
+                  Minor
+                </div>
+              </Button>
+              <Button
+                variant="bordered"
+                className={`h-full ${getButtonStyle("Major")}`}
+                onClick={() => handleSelect("Major")}
+              >
+                <div className="flex flex-col min-w-32 px-2 font-sm rounded-3xl justify-center items-center h-full">
+                  <img
+                    width="25"
+                    height="auto"
+                    src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/fault_icon_major.webp"
+                    alt="Major"
+                  />
+                  Major
+                </div>
+              </Button>
+              <Button
+                variant="bordered"
+                className={`h-full ${getButtonStyle("Critical")}`}
+                onClick={() => handleSelect("Critical")}
+              >
+                <div className="flex flex-col min-w-32 px-2 font-sm rounded-3xl justify-center items-center h-full">
+                  <img
+                    width="25"
+                    height="auto"
+                    src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/fault_icon_critical.webp"
+                    alt="Critical"
+                  />
+                  Critical
+                </div>
+              </Button>
+            </ButtonGroup>
+          </div>
+        </div>
+
+        {/* Image Upload */}
+        <div className="flex flex-col">
+          <span className="font-semibold text-sm mb-2">Attach Image</span>
+          <div className="flex border rounded-lg overflow-hidden">
+            {/* Camera upload */}
+            <div className="flex justify-center p-2 w-1/5 bg-gray-100">
+              <img
+                src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/camera_icon.webp"
+                alt="Camera Icon"
+                className="h-10 w-10"
+              />
+            </div>
+
+            {/* Drag and drop */}
+            <div className="flex border-l w-4/5 p-3 bg-white">
+              <div
+                className="w-full flex justify-center items-center"
+                {...getRootProps()}
+              >
+                <input {...getInputProps()} />
+                {isDragActive ? (
+                  <div>
+                    <svg
+                      xmlns="https://www.w3.org/2000/svg"
+                      viewBox="0 0 24 24"
+                      height="40"
+                      width="40"
+                    >
+                      <path
+                        d="M1 14.5C1 12.1716 2.22429 10.1291 4.34315 8.65685C6.46201 7.18458 9.03799 6.5 12 6.5C14.962 6.5 17.538 7.18458 19.6569 8.65685C21.7757 10.1291 23 12.1716 23 14.5V15.5C23 17.8284 21.7757 19.8709 19.6569 21.3431C17.538 22.8154 14.962 23.5 12 23.5C9.03799 23.5 6.46201 22.8154 4.34315 21.3431C2.22429 19.8709 1 17.8284 1 15.5V14.5Z"
+                        stroke="#333"
+                        strokeWidth="2"
+                        strokeLinecap="round"
+                        strokeLinejoin="round"
+                      />
+                    </svg>
+                  </div>
+                ) : (
+                  <div className="text-sm text-gray-600">
+                    Drag & drop files here, or click to browse
+                  </div>
+                )}
+              </div>
+            </div>
+          </div>
+        </div>
+
+        {/* Submit Button */}
+        <Button
+          type="submit"
+          disabled={!isFormValid}
+          className={cn(
+            "m-auto w-24 px-4 py-2 font-bold rounded-3xl transition duration-300",
+            isFormValid
+              ? "bg-blue-500 text-white hover:bg-blue-600"
+              : "bg-blue-200 text-white cursor-not-allowed"
+          )}
+        >
+          Submit
+        </Button>
+      </form>
+    </div>
+  </div>
+
+  {/* Graphical Section */}
+  <div className="w-1/2 relative bg-gray-200">
+    {/* Map Section */}
+    <div>
+      <div className="absolute flex flex-col gap-y-5 bottom-10 right-5 z-30">
+        <Button
+          className="min-w-fit h-fit p-2 bg-white"
+          onClick={panToCurrentLocation}
+        >
+          <Locate size={21} />
+        </Button>
+        <Button
+          className="min-w-fit h-fit p-2 bg-white"
+          onClick={onPinClick}
+        >
+          {!isPinDropped ? <Pin size={21} /> : <PinOff size={21} />}
+        </Button>
+      </div>
+
+      {!isPinDropped && (
+        <div className="absolute top-1/2 left-1/2 translate-x-[-50%] translate-y-[-50%] z-40">
+          <CustomMarker fill="#BE0505" />
+        </div>
+      )}
+
+      <div className="absolute text-sm bottom-5 left-1/2 transform -translate-x-1/2 z-40 p-4 bg-white bg-opacity-90 border rounded-lg shadow-lg text-black text-center max-w-sm">
+        <div
+          className="flex justify-center cursor-pointer"
+          onClick={handleTooltipToggle}
+        >
+          <Info size={24} className="text-blue-500" />
+        </div>
+        {tooltipVisible && (
+          <p>
+            Click on the pin in the bottom right corner to select this
+            location as the fault address.
+          </p>
+        )}
+      </div>
+
+      <div className="w-full h-full relative" ref={mapContainer}></div>
+    </div>
+  </div>
+</div>
+
+
+      {/* Mobile View */}
+      <div className="block sm:hidden flex flex-col justify-center items-center h-full w-full px-4 rounded-3xl overflow-hidden">
+        <ToastContainer />
+
+        <div className="w-full max-w-screen-md p-6 bg-white flex flex-col rounded-3xl justify-center overflow-y-auto">
           {/* Heading */}
           <h2 className="text-2xl text-center font-bold mt-2 mb-6">
             Report a Fault
-          </h2>{" "}
+          </h2>
           <form
             ref={formRef}
             onSubmit={handleSubmit}
-            className="flex flex-col gap-y-8"
+            className="flex flex-col gap-y-4"
           >
             {/* Fault Type */}
             {faultTypes.length > 0 ? (
@@ -319,101 +570,75 @@ const CreateTicketComp: React.FC<Props> = ({ className, useMapboxProp }) => {
               </div>
             </div>
 
+            {/* Map Toggle */}
+            <div
+              className="flex items-center justify-center mt-4 cursor-pointer"
+              onClick={() => setIsMapOpen(true)}
+            >
+              <FiMap size={24} className="text-blue-500" />
+              <FiArrowRight size={24} className="text-blue-500 ml-2" />
+            </div>
+
             {/* Fault Severity */}
             <div>
               <span className="font-semibold text-sm">Fault Severity</span>
-              <div className="flex h-[3.5rem] justify-center">
+              <div className="flex h-[2.5rem] justify-center">
                 <ButtonGroup aria-label="Basic example" className="flex h-full">
                   <Button
                     variant="bordered"
-                    className={`h-full ${getButtonStyle("Minor")}`}
-                    onClick={() => handleSelect("Minor")}
+                    className={`h-full ${
+                      selectedFault === "Minor"
+                        ? "border-blue-500 border-2"
+                        : ""
+                    }`}
+                    onClick={() => setSelectedFault("Minor")}
                   >
-                    <div className="flex flex-col min-w-32 px-2 font-sm rounded-3xl justify-center items-center h-full">
+                    <div className="flex flex-col min-w-24 px-1 font-sm rounded-2xl justify-center items-center h-full">
                       <img
-                        width="25"
+                        width="20"
                         height="auto"
                         src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/fault_icon_minor.webp"
                         alt="Minor"
                       />
-                      Minor
                     </div>
                   </Button>
                   <Button
                     variant="bordered"
-                    className={`h-full ${getButtonStyle("Major")}`}
-                    onClick={() => handleSelect("Major")}
+                    className={`h-full ${
+                      selectedFault === "Major"
+                        ? "border-blue-500 border-2"
+                        : ""
+                    }`}
+                    onClick={() => setSelectedFault("Major")}
                   >
-                    <div className="flex flex-col min-w-32 px-2 font-sm rounded-3xl  justify-center items-center h-full">
+                    <div className="flex flex-col min-w-24 px-1 font-sm rounded-2xl justify-center items-center h-full">
                       <img
-                        width="25"
+                        width="20"
                         height="auto"
                         src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/fault_icon_major.webp"
                         alt="Major"
                       />
-                      Major
                     </div>
                   </Button>
                   <Button
                     variant="bordered"
-                    className={`h-full ${getButtonStyle("Critical")}`}
-                    onClick={() => handleSelect("Critical")}
+                    className={`h-full ${
+                      selectedFault === "Critical"
+                        ? "border-blue-500 border-2"
+                        : ""
+                    }`}
+                    onClick={() => setSelectedFault("Critical")}
                   >
-                    <div className="flex flex-col min-w-32 px-2 font-sm rounded-3xl justify-center items-center h-full">
+                    <div className="flex flex-col min-w-24 px-1 font-sm rounded-2xl justify-center items-center h-full">
                       <img
-                        width="25"
+                        width="20"
                         height="auto"
                         src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/fault_icon_critical.webp"
                         alt="Critical"
                       />
-                      Critical
                     </div>
                   </Button>
                 </ButtonGroup>
-              </div>
-            </div>
-
-            {/* Image Upload */}
-            <div className=" flex flex-col">
-              <span className="font-semibold text-sm">Attach Image</span>
-              <div className="flex border">
-                {/* Camera upload */}
-                <div className="flex justify-center p-2 w-1/4">
-                  <img
-                    src="https://mycity-storage-bucket.s3.eu-west-1.amazonaws.com/resources/camera_icon.webp"
-                    alt="Camera Icon"
-                  />
-                </div>
-
-                {/* Drag and drop */}
-                <div className="flex border w-3/4">
-                  <div
-                    className="border w-full flex justify-center items-center"
-                    {...getRootProps()}
-                  >
-                    <input {...getInputProps()} />
-                    {isDragActive ? (
-                      <div>
-                        <svg
-                          xmlns="https://www.w3.org/2000/svg"
-                          viewBox="0 0 24 24"
-                          height="50"
-                          width="50"
-                        >
-                          <path
-                            d="M1 14.5C1 12.1716 2.22429 10.1291 4.34315 8.65685C6.46201 7.18458 9.03799 6.5 12 6.5C14.962 6.5 17.538 7.18458 19.6569 8.65685C21.7757 10.1291 23 12.1716 23 14.5V15.5C23 17.8284 21.7757 19.8709 19.6569 21.3431C17.538 22.8154 14.962 23.5 12 23.5C9.03799 23.5 6.46201 22.8154 4.34315 21.3431C2.22429 19.8709 1 17.8284 1 15.5V14.5Z"
-                            stroke="#333"
-                            stroke-width="2"
-                            stroke-linecap="round"
-                            stroke-linejoin="round"
-                          />
-                        </svg>
-                      </div>
-                    ) : (
-                      <div>Drop your files here or click to browse</div>
-                    )}
-                  </div>
-                </div>
               </div>
             </div>
 
@@ -433,49 +658,69 @@ const CreateTicketComp: React.FC<Props> = ({ className, useMapboxProp }) => {
           </form>
         </div>
 
-        {/* Graphical Section */}
-        <div className="w-1/2 relative bg-gray-200">
-          {/* Map Section */}
-          <div>
-            <div className="absolute flex flex-col gap-y-5 bottom-10 right-5 z-30">
-              <Button
-                className="min-w-fit h-fit p-2 bg-white"
-                onClick={panToCurrentLocation}
-              >
-                <Locate size={21} />
-              </Button>
-              <Button
-                className="min-w-fit h-fit p-2 bg-white"
-                onClick={onPinClick}
-              >
-                {!isPinDropped ? <Pin size={21} /> : <PinOff size={21} />}
-              </Button>
-            </div>
-
-            {!isPinDropped && (
-              <div className="absolute top-1/2 left-1/2 translate-x-[-50%] translate-y-[-50%] z-40">
-                <CustomMarker fill="#BE0505" />
-              </div>
-            )}
-
-            <div className="absolute text-sm bottom-5 left-1/2 transform -translate-x-1/2 z-40 p-4 bg-white bg-opacity-90 border rounded-lg shadow-lg text-black text-center max-w-sm">
+        {/* Map Modal */}
+        {isMapOpen && (
+          <div className="fixed inset-0 bg-black bg-opacity-75 z-50 flex flex-col items-center justify-center">
+            <div className="w-full h-full relative bg-white">
+              {/* Button to close the map and return to the form view */}
               <div
-                className="flex justify-center cursor-pointer"
-                onClick={handleTooltipToggle}
+                className="absolute top-4 left-4 z-50 text-black cursor-pointer bg-white p-2 rounded-full shadow-lg flex items-center"
+                onClick={() => setIsMapOpen(false)}
               >
-                <Info size={24} className="text-blue-500" />
+                <FiArrowRight
+                  size={24}
+                  className="text-blue-500 rotate-180 mr-2"
+                />{" "}
+                Back to Form
               </div>
-              {tooltipVisible && (
-                <p>
-                  Click on the pin in the bottom right corner to select this
-                  location as the fault address.
-                </p>
-              )}
-            </div>
 
-            <div className="w-full h-full relative" ref={mapContainer}></div>
+              {/* Button to pin current location */}
+              <div
+                className="absolute bottom-10 left-4 z-50 text-black cursor-pointer bg-white p-2 rounded-full shadow-lg"
+                onClick={() => {
+                  navigator.geolocation.getCurrentPosition(
+                    (position) => {
+                      const { latitude, longitude } = position.coords;
+
+                      // Create a partial PKResult object with necessary fields
+                      const location: Partial<PKResult> = {
+                        lng: longitude,
+                        lat: latitude,
+                        name: "Current Location", // Use a placeholder or actual name if available
+                        city: "",
+                        county: "",
+                        administrative: "",
+                      };
+
+                      panMapTo(longitude, latitude);
+                      dropPin(true, location as PKResult); // Cast to PKResult to satisfy TypeScript
+                    },
+                    (error) => {
+                      console.error("Error fetching location:", error);
+                      toast.error("Could not get current location.");
+                    }
+                  );
+                }}
+              >
+                <FiMapPin size={24} className="text-black" />
+              </div>
+
+              {/* Info Box */}
+              <div className="absolute bottom-5 left-1/2 transform -translate-x-1/2 z-50 p-4 bg-white bg-opacity-90 border rounded-lg shadow-lg text-black text-center max-w-sm">
+                <p>
+                  Click on the map to place a pin, or use the button on the left
+                  to pin your current location.
+                </p>
+              </div>
+
+              {/* Map Container */}
+              <div
+                className="w-full h-full relative z-40"
+                ref={mapContainer}
+              ></div>
+            </div>
           </div>
-        </div>
+        )}
       </div>
     </div>
   );
