+++ frontend/src/components/Navbar/NavbarGuest.tsx
@@ -60,9 +60,9 @@ export default function NavbarGuest() {
       </div>
 
       {/* Mobile View */}
-      <div className="block sm:hidden bottom-0 w-full h-20 bg-black bg-opacity-70 fixed flex items-center justify-center">
-        <nav className="z-40 fixed w-full p-0 flex items-center justify-center">
-          <div className="flex-initial text-[0.95rem] gap-9 flex me-5 justify-center">
+      <div className="block sm:hidden">
+        <nav className="bottom-0 w-full h-20 bg-black bg-opacity-70 fixed flex items-center justify-center">
+          <div className="flex-initial text-[0.95rem] flex w-full justify-around px-4">
             <Link href="/" passHref>
               <div className="text-white cursor-pointer transform hover:scale-105 transition-transform duration-200">
                 <div className="flex flex-col gap-1 items-center">
