+++ frontend/src/components/Navbar/NavbarUser.tsx
@@ -172,52 +172,51 @@ export default function NavbarUser({ unreadNotifications = 0 }) {
       {/* Mobile View */}
 
       <div className="block sm:hidden">
-        {/* Bottom Bar */}
-        <div className="bottom-0 w-full h-20 bg-black bg-opacity-70 fixed flex items-center justify-center">
-          <nav className="z-51 fixed w-full p-0 flex items-center justify-center">
-            <div className="flex-initial text-[0.95rem] flex me-5 gap-9 items-center">
-              <Link href="/dashboard/citizen" passHref>
-                <div className={getNavItemClass("/dashboard/citizen")}>
-                  <div className="flex flex-col gap-1 items-center">
-                    <Home size={50} />
-                  </div>
-                </div>
-              </Link>
-
-              <Link href="/create-ticket/citizen" passHref>
-                <div className={getNavItemClass("/create-ticket/citizen")}>
-                  <div className="flex flex-col gap-1 items-center">
-                    <PlusCircle size={50} />
-                  </div>
-                </div>
-              </Link>
-
-              <Link href="/notifications/citizen" passHref>
-                <div className={getNavItemClass("/notifications/citizen")}>
-                  <div className="relative flex flex-col gap-1 items-center">
-                    <Bell size={50} />
-                    {unreadNotifications > 0 && (
-                      <div className="absolute top-0 right-0 h-5 w-5 bg-blue-500 text-white text-[10px] font-bold flex items-center justify-center rounded-full z-100">
-                        {unreadNotifications}
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </Link>
-
-              <Link href="/search/citizen" passHref>
-                <div className={getNavItemClass("/search/citizen")}>
-                  <div className="flex flex-col gap-1 items-center">
-                    <Search size={50} />
-                  </div>
-                </div>
-              </Link>
-
-              
+  {/* Bottom Bar */}
+  <div className="bottom-0 w-full h-20 bg-black bg-opacity-70 fixed flex items-center justify-center">
+    <nav className="z-51 fixed w-full p-0 flex items-center justify-between">
+      <div className="flex w-full text-[0.95rem] items-center justify-between px-6">
+        <Link href="/dashboard/citizen" passHref>
+          <div className={getNavItemClass("/dashboard/citizen")}>
+            <div className="flex flex-col gap-1 items-center">
+              <Home size={50} />
             </div>
-          </nav>
-        </div>
+          </div>
+        </Link>
+
+        <Link href="/create-ticket/citizen" passHref>
+          <div className={getNavItemClass("/create-ticket/citizen")}>
+            <div className="flex flex-col gap-1 items-center">
+              <PlusCircle size={50} />
+            </div>
+          </div>
+        </Link>
+
+        <Link href="/notifications/citizen" passHref>
+          <div className={getNavItemClass("/notifications/citizen")}>
+            <div className="relative flex flex-col gap-1 items-center">
+              <Bell size={50} />
+              {unreadNotifications > 0 && (
+                <div className="absolute top-0 right-0 h-5 w-5 bg-blue-500 text-white text-[10px] font-bold flex items-center justify-center rounded-full z-100">
+                  {unreadNotifications}
+                </div>
+              )}
+            </div>
+          </div>
+        </Link>
+
+        <Link href="/search/citizen" passHref>
+          <div className={getNavItemClass("/search/citizen")}>
+            <div className="flex flex-col gap-1 items-center">
+              <Search size={50} />
+            </div>
+          </div>
+        </Link>
       </div>
+    </nav>
+  </div>
+</div>
+
     </div>
   );
 }
